16:53 17.05.2024
przez wstawieniem policy

+++ app/Http/Controllers/FishController.php
@@ -49,7 +49,36 @@ class FishController extends Controller
 
     public function destroy($id){
         $fish = Fish::findORFail($id);
+
+        if ($fish->relatedRecordsExist()) {
+            return back()->with('error', 'Nie można usunąć ryby, ponieważ istnieją powiązane rekordy w innych tabelach.');
+        }
+
         $fish->delete();
         return redirect()->route('fish.index');
+    }
+
+    public function store(Request $request)
+    {
+        // Walidacja danych wejściowych
+        $validatedData = $request->validate([
+            'name' => 'required|string|max:255',
+            'species' => 'required|string|max:255',
+            'description' => 'required|string',
+            'image' => 'nullable|string',
+        ]);
+
+        // Utwórz nowy obiekt ryby na podstawie danych wejściowych
+        $fish = new Fish();
+        $fish->name = $validatedData['name'];
+        $fish->species = $validatedData['species'];
+        $fish->description = $validatedData['description'];
+        $fish->image = $validatedData['image']; // Możesz dodać logikę przetwarzania obrazów, jeśli jest to wymagane
+
+        // Zapisz rybę do bazy danych
+        $fish->save();
+
+        // Przekieruj użytkownika gdziekolwiek chcesz, np. do listy wszystkich ryb
+        return redirect()->route('fish.index')->with('success', 'Ryba została pomyślnie dodana.');
     }
 }
